\name{plotly-package}
\alias{plotly-package}
\docType{package}
\title{A R API for plot.ly}
\description{
plot.ly is a browser-based data visualization tool that creates interactive and publication quality figures. This API allows R users to generate plot.ly graphs from their desktop R environment.
Here is an example of an interactive graph made from the R API: https://plot.ly/	~chris/407/
Documentation and examples at https://plot.ly/API
}
\details{
Documentation and examples at https://plot.ly/API
\tabular{ll}{
Package: \tab plotly\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2013-06-12\cr
License: \tab MIT\cr
}
}
\author{
Chris P <chris@plot.ly>}
\references{
Documentation and examples at https://plot.ly/API
}

\keyword{ package }
\seealso{
}
\examples{
# Up-to-date documentation and examples at https://plot.ly/API
# These examples may not be up-to-date

}
